+++ examples/postman_echo/request_methods/request_with_functions.yml
@@ -26,8 +26,8 @@ teststeps:
     validate:
         - eq: ["status_code", 200]
         - eq: ["body.args.foo1", "session_bar1"]
+        - eq: ["body.args.sum_v", 3]
         - eq: ["body.args.foo2", "session_bar2"]
-        - eq: ["body.args.sum_v", "3"]
 -
     name: post raw text
     variables:
