+++ examples/postman_echo/request_methods/request_with_functions_test.py
@@ -1,5 +1,5 @@
-from httprunner.v3.runner import TestCaseRunner
-from httprunner.v3.schema import TestsConfig, TestStep
+from httprunner.runner import TestCaseRunner
+from httprunner.schema import TestsConfig, TestStep
 from examples.postman_echo import debugtalk
