+++ httprunner/loader/__init__.py
@@ -8,16 +8,14 @@ HttpRunner loader
 
 """
 
-from httprunner.loader.check import is_test_path, is_test_content, JsonSchemaChecker
+from httprunner.loader.buildup import load_cases, load_project_data
+from httprunner.loader.check import is_test_path
+from httprunner.loader.load import load_csv_file, load_builtin_functions
 from httprunner.loader.locate import get_project_working_directory as get_pwd, \
     init_project_working_directory as init_pwd
-from httprunner.loader.load import load_csv_file, load_builtin_functions
-from httprunner.loader.buildup import load_cases, load_project_data
 
 __all__ = [
     "is_test_path",
-    "is_test_content",
-    "JsonSchemaChecker",
     "get_pwd",
     "init_pwd",
     "load_csv_file",
