--- httprunner/schema/__init__.py
@@ -1,3 +1,0 @@
-from .api import Api
-from .testcase import ProjectMeta, TestCase
-from .testsuite import TestSuite
